+++ products/data.go
@@ -1,4 +1,4 @@
-package main
+package products
 
 // Request
 type ProductRequest struct {
